ap1302: Add support for dual sensors
According to the datasheet the value written in the PREVIEW_WIDTH
register must be at most the width of the image sensor.

When using a dual sensors setup the total resolution is 2 times the
width by the height of one sensor since the images are side by side.
Thus to retrieve the width of one sensor we need to divide the total
width by the number of sensors.

Signed-off-by: Florian Rebaudo <frebaudo@witekio.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>